added a 1d vector x 1d vector element wise multiplication example
