+++ example/ck_tile/21_multiply_vector/CMakeLists.txt
@@ -1,9 +1,9 @@
-set(EXAMPLE_REDUCE "add_vector")
+set(EXAMPLE_REDUCE "multiply_vector")
 # not using add_example_executable() to add this target, since we don't want this to have
 # to be included in "make all/install/check"
 message("adding example ${EXAMPLE_REDUCE}")
 
-add_executable(${EXAMPLE_REDUCE} EXCLUDE_FROM_ALL add_vector.cpp)
+add_executable(${EXAMPLE_REDUCE} EXCLUDE_FROM_ALL multiply_vector.cpp)
 target_include_directories(${EXAMPLE_REDUCE} PRIVATE ${CMAKE_CURRENT_LIST_DIR})
 set(EXAMPLE_REDUCE_COMPILE_OPTIONS)
