+++ example/ck_tile/21_multiply_vector/multiply_vector.hpp
@@ -1,0 +1,135 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2018-2024, Advanced Micro Devices, Inc. All rights reserved.
+
+#pragma once
+
+#include "ck_tile/core.hpp"
+#include "ck_tile/ops/common.hpp"
+
+namespace ck_tile {
+
+template <typename BlockWarps, // num warps along seq<M, N>
+          typename BlockTile,  // block size, seq<M, N>
+          typename WarpTile,   // warp size, seq<M, N>
+          typename Vector>     // contiguous pixels(vector size) along seq<M, N>
+struct MultiplyShape
+{
+    static constexpr index_t Block_M = BlockTile::at(number<0>{});
+    //static constexpr index_t Block_N = BlockTile::at(number<1>{});
+
+    static constexpr index_t Warp_M = WarpTile::at(number<0>{});
+    //static constexpr index_t Warp_N = WarpTile::at(number<1>{});
+
+    static constexpr index_t Vector_M = Vector::at(number<0>{});
+    //static constexpr index_t Vector_N = Vector::at(number<1>{});
+
+    // TODO: calculate it from BLOCK_M, BLOCK_N, WARP_M, WARP_N
+    static constexpr index_t WarpPerBlock_M = BlockWarps::at(number<0>{});
+    //static constexpr index_t WarpPerBlock_N = BlockWarps::at(number<1>{});
+
+    static constexpr index_t ThreadPerWarp_M = Warp_M / Vector_M;
+    //static constexpr index_t ThreadPerWarp_N = Warp_N / Vector_N;
+
+    static constexpr index_t Repeat_M = Block_M / (WarpPerBlock_M * Warp_M);
+    //static constexpr index_t Repeat_N = Block_N / (WarpPerBlock_N * Warp_N);
+
+    static constexpr index_t BlockSize =
+        warpSize * reduce_on_sequence(BlockWarps{}, multiplies{}, number<1>{});
+};
+
+template <typename XDataType_,
+          typename ComputeDataType_,
+          typename YDataType_,
+          typename BlockShape_>
+struct MultiplyProblem
+{
+    using XDataType       = remove_cvref_t<XDataType_>;
+    using ComputeDataType = remove_cvref_t<ComputeDataType_>;
+    using YDataType       = remove_cvref_t<YDataType_>;
+    using BlockShape      = remove_cvref_t<BlockShape_>;
+};
+
+struct MultiplyDefaultPolicy
+{   
+    // Distribution for 1D vector operations
+    template <typename Problem>
+    CK_TILE_DEVICE static constexpr auto MakeXBlockTileDistribution()
+    {
+        using S = typename Problem::BlockShape;
+        return make_static_tile_distribution(
+            tile_distribution_encoding<
+                sequence<>,
+                // Simplified distribution focusing only on M dimension (vector)
+                tuple<sequence<S::Repeat_M, S::WarpPerBlock_M, S::ThreadPerWarp_M, S::Vector_M>>,
+                tuple<sequence<1>, sequence<1>>,
+                tuple<sequence<1>, sequence<2>>,
+                sequence<1, 1>,
+                sequence<0, 3>>{});
+    }
+};
+
+template <typename Problem_, typename Policy_ = MultiplyDefaultPolicy>
+struct MultiplyVector
+{
+    using Problem = ck_tile::remove_cvref_t<Problem_>;
+    using Policy  = ck_tile::remove_cvref_t<Policy_>;
+
+    using XDataType       = ck_tile::remove_cvref_t<typename Problem::XDataType>;
+    using ComputeDataType = ck_tile::remove_cvref_t<typename Problem::ComputeDataType>;
+    using YDataType       = ck_tile::remove_cvref_t<typename Problem::YDataType>;
+
+    CK_TILE_DEVICE void operator()(const XDataType* p_x_a, const XDataType* p_x_b, YDataType* p_y, index_t M) const
+    {
+        using S = typename Problem::BlockShape;
+
+        // Create 1D tensor views for the vectors
+        const auto x_m_a = make_naive_tensor_view<address_space_enum::global>(
+            p_x_a, make_tuple(M), make_tuple(1), number<1>{});
+        
+        const auto x_m_b = make_naive_tensor_view<address_space_enum::global>(
+            p_x_b, make_tuple(M), make_tuple(1), number<1>{});
+
+        const auto y_m = make_naive_tensor_view<address_space_enum::global>(
+            p_y, make_tuple(M), make_tuple(1), number<1>{});
+
+        // Calculate starting index for this block
+        const auto iM = get_block_id() * S::Block_M;
+
+        // Create tile windows for each vector
+        // x_window_a and x_window_b are tile windows
+        // they define a mapping from the global tensor view to the tile view
+        auto x_window_a = make_tile_window(x_m_a, // global tensor view
+                                         make_tuple(number<S::Block_M>{}), // tile dimensions
+                                         {iM}, // starting index
+                                         Policy::template MakeXBlockTileDistribution<Problem>()); // distribution
+
+        auto x_window_b = make_tile_window(x_m_b,
+                                         make_tuple(number<S::Block_M>{}),
+                                         {iM},
+                                         Policy::template MakeXBlockTileDistribution<Problem>());
+
+        auto y_window = make_tile_window(y_m,
+                                         make_tuple(number<S::Block_M>{}),
+                                         {iM},
+                                         Policy::template MakeXBlockTileDistribution<Problem>());
+
+        // Load tile data
+        const auto xa = load_tile(x_window_a); // load tile data from global tensor view, load from where? what? how many? logical memory layout? all are defined in x_window_a
+        const auto xb = load_tile(x_window_b);
+        auto y_compute = load_tile(y_window);
+
+        // Process the vector multiplication
+        constexpr auto spans = decltype(xa)::get_distributed_spans(); // shape of the tile
+        sweep_tile_span(spans[number<0>{}], [&](auto idx) { // iterate over the tile
+            const auto tile_idx = make_tuple(idx);
+            const auto a_val = type_convert<ComputeDataType>(xa[tile_idx]);
+            const auto b_val = type_convert<ComputeDataType>(xb[tile_idx]);
+            y_compute(tile_idx) = a_val * b_val;
+        });
+
+        // Store results
+        store_tile(y_window, cast_tile<YDataType>(y_compute)); // store the result back to global tensor view
+    }
+};
+
+} // namespace ck_tile
