+++ example/ck_tile/CMakeLists.txt
@@ -18,4 +18,4 @@ add_subdirectory(15_fused_moe)
 add_subdirectory(16_batched_gemm)
 add_subdirectory(17_grouped_gemm)
 add_subdirectory(20_add)
-add_subdirectory(21_add_vector)
+add_subdirectory(21_multiply_vector)
