+++ include/ck_tile/host.hpp
@@ -35,3 +35,4 @@
 #include "ck_tile/host/stream_config.hpp"
 #include "ck_tile/host/timer.hpp"
 #include "ck_tile/host/reference/reference_add.hpp"
+#include "ck_tile/host/reference/reference_multiply_vector.hpp"
